# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "UserPost demo API",
        "description": "Node.js and Express.js Application API",
        "license": {
            "name": "Post",
            "url": "https://post.com/"
        }
    },
...
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "in": "cookie",
            "name": "token"
        }
    },
    "security": [
        {
            "jwt": []
        }
    ],
    "paths": {
        "/insertuser": {
            "post": {
              "tags": ["User"],
              "summary": "Signup API",
              "security" : [],
              "parameters": [
                {
                  "name": "user",
                  "in": "body",
                  "description": "Create new account",
                  "schema": {
                    "allOf":[
                      {
                        "$ref": "#/definitions/SignUp"
                      }
                    ]
                  }
                }
              ],
              "produces": ["application/json"],
              "responses": {
                "200": {
                  "description": "Create new Account",
                  "schema": {
                    "allOf": [
                      {
                        "type": "object",
                        "properties": {
                          "status"  : { "type": "boolean"},
                          "data"    : { 
                            "type": "object",
                            "properties": {
                              "_id" : { "type" : "string"},
                              "fullname" : { "type" : "string"},
                              "username" : { "type" : "string"},
                              "email"    : { "type" : "string"},
                              "password" : { "type" : "string"}
                            }
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
        },
        "/loginuser":{
          "post":{
            "tags":["User"],
              "summary":"Login API",
              "security" : [],
              "parameters": [
                { 
                  "name": "user",
                  "in": "body",
                  "description": "Login account",
                  "schema": {
                    "type": "object",
                    "required"  : ["email", "password"],
                    "properties": {
                      "email"     : { "type": "string"},
                     "password"  : { "type": "string"}
                    }
                  }
                }
              ],
              "produces":["application/json"],
              "responses": {
                "200": {
                  "description": "Login account",
                  "schema": {
                    "allOf": [
                      {
                        "$ref": "#/definitions/LoginResponse"
                      }
                    ]
                  }
                }
              }
            }
        },
...
        "/updateuser/{id}":{
          "parameters":[
            {
              "name":"id",
              "in":"path",
              "required":true,
              "description":"ID of User",
              "type":"string"
            }
          ],
          "post":{
            "tags":["User"],
            "summary":"Update UserProfile",
            "parameters":[
              {
                "in"          : "body",
                "name"        : "user",
                "description" : "Update user Profile",
                "schema":{
                  "allOf" :[
                    {
                      "$ref":"#/definitions/Profile"
                    },
                    {
                      "type" : "object",
                      "required" : ["fullname","username","email"]
                    }
                  ]
                }
              }
            ],
            "produces":["application/json"],
            "responses":{
              "200":{
                "description":"Update UserProfile",
                "schema":{
                  "allOf":[
                    {
                      "type":"object",
                      "properties":{
                        "status":{"type" : "boolean"},
                        "data":{
                          "allof":[
                            {"$ref": "#/definitions/Profile"},
                            {
                              "type":"object",
                              "properties":{
                                "_id"  : { "type" : "string"},
                                "user" : { "type" : "string"}
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
...
        "/mypost/{id}":{
          "parameters":[
            {
              "name":"id",
              "in":"path",
              "required":true,
              "description":"ID of Post",
              "type":"string"
            }
          ],
          "get":{
            "tags":["Post"],
            "summary":"Get Post By id",
            "responses":{
              "200":{
                "description":"Get Post by Id",
                "schema":{
                  "allOf":[
                    {
                      "type":"object",
                      "properties":{
                        "status":{"type" : "boolean"},
                        "data":{
                          "allof":[
                            {"$ref":"#/definitions/Post"},
                            {
                              "type":"object",
                              "user":{
                                "type":"object",
                                "properties":{
                                  "_id":{"type":"string"},
                                  "fullname":{"type":"string"},
                                  "username":{"type":"string"}
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
...
        "/admin/makeadmin/{id}":{
          "parameters":[
            {
              "name":"id",
              "in":"path",
              "required":true,
              "description":"Make Admin",
              "type":"string"
            }
          ],
          "get":{
            "tags":["Admin"],
            "summary":"Make Admin",
            "responses":{
              "200":{
                "description":"Make Admin",
                "schema":{
                  "allOf":[
                    {
                      "type":"object",
                      "properties":{
                        "status":{"type" : "boolean"},
                        "data":{
                          "allof":[
                            {"$ref":"#/definitions/Profile"},
                            {
                              "type":"object",
                              "user":{
                                "type":"object",
                                "properties":{
                                  "_id":{"type":"string"},
                                  "fullname":{"type":"string"},
                                  "username":{"type":"string"}
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "/admin/removeadmin/{id}":{
          "parameters":[
            {
              "name":"id",
              "in":"path",
              "required":true,
              "description":"Remove Admin",
              "type":"string"
            }
          ],
          "get":{
            "tags":["Admin"],
            "summary":"Remove Admin",
            "responses":{
              "200":{
                "description":"Remove Admin",
                "schema":{
                  "allOf":[
                    {
                      "type":"object",
                      "properties":{
                        "status":{"type" : "boolean"},
                        "data":{
                          "allof":[
                            {"$ref":"#/definitions/Profile"},
                            {
                              "type":"object",
                              "user":{
                                "type":"object",
                                "properties":{
                                  "_id":{"type":"string"},
                                  "fullname":{"type":"string"},
                                  "username":{"type":"string"}
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
    },
    "definitions": {
      "SignUp": {
        "required"  : ["fullname", "username", "email", "password"],
        "properties": {
          "fullname" : { "type"  : "string" },
          "username"  : { "type"  : "string" },
          "email"     : { "type"  : "string" },
          "password"  : { "type"  : "string" }
        }
      },
      "Profile": {
        "required"  : ["fullname", "username", "email"],
        "properties": {
          "fullname" : { "type"  : "string" },
          "username"  : { "type"  : "string" },
          "email"     : { "type"  : "string" }
        }
      },
      "LoginResponse":{
        "properties":{
          "fullname" : { "type" : "string"},
          "username" : { "type" : "string" },
          "email"    : { "type" : "string" },
          "token"    : { "type" : "string" },
          "role"     : { 
                        "type"           : "object",
                         "properties"    : {
                            "_id"        : { "type" : "string"},
                            "rolename"   : { "type" : "string"}
                         }                         
                        },
          "_id"       : { "type" : "string"}
        }
      },
      "Post": {
        "properties" : {
          "title"         :{ "type" : "string"},
          "description"   :{ "type" : "string"}
        }
      },
...
    }
}